+++ g5/car_software/libat90/io.h
@@ -1,6 +1,8 @@
 #ifndef IO_H
 #define IO_H
 
+#include <stdint.h>
+
 #define HIGH 	1
 #define LOW 	0
 
@@ -10,10 +12,10 @@ enum io_pinmode_t {
 	INPUT_PULLUP
 };
 
-int digitalRead(int inputPinRegister, int pin);
-void digitalWrite(volatile int *port, int pin, int value);
+int digitalRead(uint8_t inputPinRegister, int pin);
+void digitalWrite(volatile uint8_t *port, int pin, int value);
 
-int pinMode(volatile int *port, int pin, enum io_pinmode_t mode);
+int pinMode(volatile uint8_t *port, int pin, enum io_pinmode_t mode);
 
 
 #endif /* IO_H */
